+++ cli/integrity.ts
@@ -45,7 +45,7 @@ export async function checkIntegrity(lock?: "check" | "update" | "ignore") {
   }
 
   if (!lock) {
-    lock = process.env["CI"] ? "check" : "ignore";
+    lock = process.env["CI"] ? "check" : "update";
   }
 
   if (lock === "update") {
@@ -197,7 +197,13 @@ export async function updateLockFile() {
 
   let rootDir = getRootDir();
   let lockFile = path.join(rootDir, "mops.lock");
+  let isNew = !fs.existsSync(lockFile);
   fs.writeFileSync(lockFile, JSON.stringify(lockFileJson, null, 2));
+  if (isNew) {
+    console.log("mops.lock created.");
+    console.log("  Applications: commit this file.");
+    console.log("  Libraries: add mops.lock to .gitignore.");
+  }
 }
 
 // compare hashes of local files with hashes from the lock file